requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLAlchemy for production DB support)
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.0

# WebSocket support
websockets==12.0

# HTTP client for webhooks
httpx==0.25.1

# Environment variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0

psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# AWS SDK (for deployment)
boto3==1.29.7
botocore==1.32.7

# Langchain for RAG (optional - uncomment if using real AI)
# langchain==0.1.0
# openai==1.3.0
# chromadb==0.4.18